pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "QSnake", email = "jester36883688@gmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
